Bug fixed, little upgrades
User input. 3 new options based on user input: deleting file, renaming file, opening or selecting after finding. No mulit-thread support now. couse of bugs.

+++ Source.cpp
@@ -2,10 +2,14 @@
 #include <iostream>
 #include <string>
 #include <thread>
+#include <cstdio>
 
 using namespace std;
 
-bool checkstr(string str, string temp) {
+int op = 0;
+int API = 0;
+char newfile[100];
+bool checkstr(string str, string temp, int wtd=0) {
 	//cout << str<<endl;
 	//int len = str.length()-1;
 	//if ((str[len] == 't') && (str[len - 1] == 'x') && (str[len - 2] == 't'))return true; //проверка на расширение
@@ -13,19 +17,18 @@ bool checkstr(string str, string temp) {
 	return false;
 }
 
-void ff(const TCHAR *curDir, string temp)
+bool ff(const TCHAR *curDir, string temp)
 {
-	Sleep(1);
 	WIN32_FIND_DATA FindFileData;
 	HANDLE hFind;
-	TCHAR cd[MAX_PATH], res[MAX_PATH];
+	TCHAR cd[MAX_PATH], res[MAX_PATH]; 
+	char filen[MAX_PATH];
 	strcpy_s(cd, MAX_PATH, curDir);
 	strcat_s(cd, MAX_PATH, "\\*.*");
 	hFind = FindFirstFile(cd, &FindFileData);
 	BOOL r = hFind != INVALID_HANDLE_VALUE;
 	while (r)
 	{
-		Sleep(1);
 		r = FindNextFile(hFind, &FindFileData);
 		if (FindFileData.cFileName[0] != '.')
 			try
@@ -33,17 +36,30 @@ void ff(const TCHAR *curDir, string temp)
 			string filename = FindFileData.cFileName;
 			strcpy_s(cd, MAX_PATH, curDir);
 			strcat_s(cd, MAX_PATH, "\\");
+			strcpy_s(filen, MAX_PATH, cd);
 			strcat_s(cd, MAX_PATH, FindFileData.cFileName);
 			CharToOem(cd, res);
-		if (checkstr(filename,temp)) { ///проверяет в функции только конкретное расширение
+			if (checkstr(filename, temp)) { ///проверяет в функции только конкретное расширение
 				cout << "Файл найден! Его путь: " << res << endl;
-				if (remove(res) == 0)cout << "File deleted!\n";
-				else RemoveDirectory(res);
-				r = false;
-				break;
+
+				if (op == 1) {
+					if (remove(res) == 0)cout << "File deleted!\n";
+					else RemoveDirectory(res);
+				}
+				else if (op==2) {
+					char path[100] = "";
+					if (API == 1) strcpy_s(path, "/select, ");
+					strcat_s(path, res);
+					ShellExecute(NULL, "open", "explorer.exe", path, NULL, SW_SHOWNORMAL);
+				}
+				else if (op==3) {
+					strcat_s(filen,newfile);
+					rename(res, filen);
+				}
+				return true;
 			}
 			if (FindFileData.dwFileAttributes&&FILE_ATTRIBUTE_DIRECTORY)
-				ff(cd,temp);
+				ff(cd, temp);
 		}
 		catch (...)
 		{
@@ -55,23 +71,23 @@ void ff(const TCHAR *curDir, string temp)
 
 int main()
 {
-	string temp;
-	string temp1;
 	setlocale(LC_ALL, "rus");
-	//t1.detach();
-	cout << "Введите название файла, который нужно найти и удалить\n";
+
+	string temp;
+	cout << "Введите название файла\n";
 	cin >> temp;
+
+	cout << "Введите номер операции (1-удалить; 2-найти или открыть; 3-переименовать; )\n";
+	cin >> op;
+	if (op == 2) { cout << "Введите 1, если хотите выделить файл без его запуска\n"; cin >> API; }
+	if (op == 3){
+		cout << "Введите новое название\n";
+		cin >> newfile;
+}
 	cout << "Searching file..\n";
-	std::thread t1(ff, "C:", temp);
-	cout << "Введите название файла, который нужно найти и удалить #2\n";
-	cin >> temp1;
-	cout << "Searching file #2..\n";
-	std::thread t2(ff, "C:", temp1);
-	t1.join();
-	//t1.join();
-	t2.join();
-	//ff("C:"); 
-	cout << "END";
+	if (ff("C:", temp))cout << "Succsess\n";
+	else { cout << "ERROR\n"; }
 	system("pause");
 	return 0;
 }
+
